restructure and working with piped_commands
