+++ Makefile
@@ -11,6 +11,7 @@ CFLAGS	= -Wall -Wextra -Werror -g
 SRC = 	main.c \
 		./utils/init_env.c \
 		./utils/cleanup.c \
+		./utils/print_struct.c \
 		./buildins/buildins.c \
 		./buildins/cd.c \
 		./buildins/env.c \
@@ -18,6 +19,7 @@ SRC = 	main.c \
 		./buildins/unset.c \
 		./buildins/export.c \
 		./buildins/echo.c \
+		./parsing/parse.c \
 		./execution/exec.c \
 
 all: $(NAME)
