+++ utils/init_env.c
@@ -33,6 +33,7 @@ void	init_env(t_env *env_vars, char **env)
 		env_vars->env_vars[i] = ft_strdup(env[i]);
 		i++;
 	}
+	env_vars->env_vars[i] = NULL;
 	env_vars->status = 0;
 }
